Start-Process => Invoke-External
Start-Process does not bind stdin/stdout/stderr to the calling
process's standard handles. While Start-Process -Wait does wait
for the child process to complete, it will under some
circumstances not pipe input/output. The correct
approach under those situations is to use the call
operator (&).

The call operator is not mockable in tests - so we introduce
a helper function, called Invoke-External. This method
is a minimal wrapper, only intended to make it easy
to create side-effect-free tests for code that wants to
call external programs.